requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
gunicorn==21.2.0
django-htmlmin==0.11.0
requests==2.31.0
netmiko==4.3.0
pyats==23.9
genie==23.9
genie[library]==23.9
slack-sdk==3.23.0
textfsm==1.1.3
paramiko==3.3.1
htmlmin==0.1.12
python-dotenv==1.0.0